chore: 1-click refactor application service
