+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationService.java
@@ -1,5 +1,5 @@
 package com.appsmith.server.services;
 
-import com.appsmith.server.services.ce.ApplicationServiceCE;
+import com.appsmith.server.services.ce_compatible.ApplicationServiceCECompatible;
 
-public interface ApplicationService extends ApplicationServiceCE {}
+public interface ApplicationService extends ApplicationServiceCECompatible {}
