+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce_compatible/ApplicationServiceCECompatible.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.services.ce_compatible;
+
+import com.appsmith.server.services.ce.ApplicationServiceCE;
+
+public interface ApplicationServiceCECompatible extends ApplicationServiceCE {}
